RHAIENG-5164: fix: keep pytest out of gha_pr_changed_files import path
GHA resolves COMPONENT_DIR via plain python3 without dev deps; move
pyfakefs symlink tests to gha_pr_changed_files_symlink_test.py.
